+++ src/main/kotlin/com/github/blarc/ai/commits/intellij/plugin/notifications/Notification.kt
@@ -43,6 +43,7 @@ data class Notification(
         fun promptTooLarge() = Notification(DEFAULT_TITLE, message = message("notifications.prompt-too-large"))
         fun unsuccessfulRequest(message: String) = Notification(message = message("notifications.unsuccessful-request", message))
         fun noCommitMessage(): Notification = Notification(message = message("notifications.no-commit-message"))
+        fun unableToSaveToken(): Notification = Notification(message = message("notifications.unable-to-save-token"))
 
     }